requirements/base.txt
pytz==2018.4  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==3.3.1  # https://github.com/evansd/whitenoise
redis>=2.10.5  # https://github.com/antirez/redis
celery==3.1.25  # pyup: <4.0  # https://github.com/celery/celery

# Django
# ------------------------------------------------------------------------------
django==2.0.5  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis
django-filter==2.0.0

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
djangorestframework-jwt==1.11.0
coreapi==2.3.3  # https://github.com/core-api/python-client
django-url-filter==0.3.5
pygments==2.2.0
markdown==2.6.11
rest-framework-generic-relations==1.2.0

# Utilities
# -------------------------------------------------------------------------------
bleach==2.1.3
django-simple-options==1.0a3

# Project specific
# -------------------------------------------------------------------------------
easy-thumbnails==2.5
django-extensions==2.0.7  # https://github.com/django-extensions/django-extensions
django-cors-headers==2.2.0
